Move chat overview from ContactInfo to the ConversationSettings screen
The summary belongs on the screen you reach by tapping the recipient at the
top of a chat (ConversationSettings — the one showing their avatar/name),
not one level deeper on ContactInfo. ContactInfo is reverted to its original
avatar+name+OdinId; ConversationSettings now loads and renders the overview.

ConversationSettingsViewModel already knows its conversationId (route param),
so it queries messages directly — no XorId, no navigation threading. Reuses
the same counting + recent-media logic and the contact_info_* strings.
