+++ homebase-common/src/commonMain/kotlin/id/homebase/core/ui/navigation/Routes.kt
@@ -41,7 +41,7 @@ sealed class Route {
 
     @Serializable
     @SerialName("contact")
-    data class ContactInfo(val odinId: String, val conversationId: String? = null) : Route()
+    data class ContactInfo(val odinId: String) : Route()
 
     @Serializable
     @SerialName("archived-conversations")
